+++ www/js/pinsController.js
@@ -2,28 +2,34 @@
 
 var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol ? "symbol" : typeof obj; };
 
-angular.module('starter.controllers').controller('PinsController', function ($scope, Auth, PinService, UserService) {
+angular.module('starter.controllers').controller('PinsController', ['$scope', '$ionicHistory', 'Auth', 'PinService', 'UserService', function ($scope, $ionicHistory, Auth, PinService, UserService) {
   $scope.Auth = Auth;
   $scope.PinService = PinService;
 
   $scope.newPin = {};
-  $scope.submittedPins = PinService.submittedPins;
-  $scope.approvedPins = PinService.approvedPins;
   $scope.addPanelExpanded = false;
 
   // any time auth state changes, add the user data to scope
-  $scope.Auth.$onAuthStateChanged(function (firebaseUser) {
-    $scope.user = firebaseUser;
-    if (!firebaseUser) {
-      $scope.admin = null;
-    }
-  });
+  // $scope.Auth.$onAuthStateChanged(function(firebaseUser) {
+  //   $scope.user = firebaseUser;
+  //   if (!firebaseUser) {
+  //     $scope.admin = null;
+  //   }
+  // });
 
   $scope.$on('$ionicView.enter', function () {
+    $ionicHistory.clearHistory();
     initialize();
   });
 
   function initialize() {
+    PinService.submittedPins.$loaded().then(function (submittedData) {
+      $scope.submittedPins = submittedData;
+      PinService.approvedPins.$loaded().then(function (approvedData) {
+        $scope.approvedPins = approvedData;
+      });
+    });
+
     // Create the autocomplete helper, and associate it with
     // an HTML text input box.
     var input = document.getElementById('placeQuery');
@@ -89,11 +95,15 @@ angular.module('starter.controllers').controller('PinsController', function ($sc
   };
 
   $scope.favorite = function (pin) {
-    $scope.PinService.addToFavorites(pin, $scope.user.uid);
+    if ($scope.user) {
+      $scope.PinService.addToFavorites(pin, $scope.user.uid);
+    }
   };
 
   $scope.unfavorite = function (pin) {
-    $scope.PinService.removeFromFavorites(pin, $scope.user.uid);
+    if ($scope.user) {
+      $scope.PinService.removeFromFavorites(pin, $scope.user.uid);
+    }
   };
 
   $scope.isFavorite = function (pin) {
@@ -123,4 +133,4 @@ angular.module('starter.controllers').controller('PinsController', function ($sc
     };
     return date.toLocaleTimeString("en-us", options);
   };
-});
+}]);
